[refactor] move `utils` to `-private` directory
