+++ addon/-private/utils/remove-from-dom.ts
