+++ addon/metrics-adapters/amplitude.js
@@ -3,8 +3,8 @@ import {
   without,
   compact,
   hasOwnProperties,
-} from 'ember-metrics/utils/object-transforms';
-import removeFromDOM from 'ember-metrics/utils/remove-from-dom';
+} from 'ember-metrics/-private/utils/object-transforms';
+import removeFromDOM from 'ember-metrics/-private/utils/remove-from-dom';
 import { assert, deprecate } from '@ember/debug';
 
 export default class AmplitudeMetricsAdapter extends BaseAdapter {
