+++ addon/metrics-adapters/intercom.js
@@ -1,6 +1,9 @@
 import { assert } from '@ember/debug';
-import { compact, without } from 'ember-metrics/utils/object-transforms';
-import removeFromDOM from 'ember-metrics/utils/remove-from-dom';
+import {
+  compact,
+  without,
+} from 'ember-metrics/-private/utils/object-transforms';
+import removeFromDOM from 'ember-metrics/-private/utils/remove-from-dom';
 import BaseAdapter from './base';
 
 export default class Intercom extends BaseAdapter {
