+++ addon/metrics-adapters/pendo.js
@@ -4,8 +4,8 @@ import {
   without,
   compact,
   hasOwnProperties,
-} from 'ember-metrics/utils/object-transforms';
-import removeFromDOM from 'ember-metrics/utils/remove-from-dom';
+} from 'ember-metrics/-private/utils/object-transforms';
+import removeFromDOM from 'ember-metrics/-private/utils/remove-from-dom';
 
 export default class Pendo extends BaseAdapter {
   toStringExtension() {
