+++ tests/unit/utils/object-transforms-test.js
@@ -2,7 +2,7 @@ import {
   without,
   compact,
   hasOwnProperties,
-} from 'ember-metrics/utils/object-transforms';
+} from 'ember-metrics/-private/utils/object-transforms';
 import { module, test } from 'qunit';
 
 module('Unit | Utility | object transforms', function () {
